+++ packages/react-ui/src/widgets/legend/LegendWrapper.js
@@ -12,6 +12,9 @@ import {
 } from '@material-ui/core';
 import { ExpandLess, ExpandMore } from '@material-ui/icons';
 import Note from './Note';
+import LayerIcon from '../../assets/LayerIcon';
+import { ToggleButton } from '@material-ui/lab';
+import OpacityControl from '../OpacityControl';
 
 const useStyles = makeStyles((theme) => ({
   legendWrapper: {
@@ -24,6 +27,9 @@ const useStyles = makeStyles((theme) => ({
   },
   attr: {
     marginBottom: theme.spacing(1)
+  },
+  layerOptionsWrapper: {
+    backgroundColor: theme.palette.grey[50]
   }
 }));
 
@@ -36,11 +42,19 @@ export default function LegendWrapper({
   note,
   attr,
   children,
+  showOpacityControl,
+  opacity,
+  onChangeOpacity,
   onChangeVisibility
 }) {
   const wrapper = createRef();
   const classes = useStyles();
   const [expanded, setExpanded] = useState(true);
+  const [isLayerOptionsExpanded, setIsLayerOptionsExpanded] = useState(false);
+
+  const handleChangeOpacity = (newOpacity) => {
+    if (onChangeOpacity) onChangeOpacity({ id, opacity: newOpacity });
+  };
 
   const handleExpandClick = () => {
     if (collapsible) {
@@ -50,6 +64,10 @@ export default function LegendWrapper({
 
   const handleChangeVisibility = () => {
     if (onChangeVisibility) onChangeVisibility({ id, visible: !visible });
+  };
+
+  const handleToggleLayerOptions = () => {
+    setIsLayerOptionsExpanded((oldState) => !oldState);
   };
 
   return (
@@ -62,6 +80,8 @@ export default function LegendWrapper({
         collapsible={collapsible}
         onExpandClick={handleExpandClick}
         onChangeVisibility={handleChangeVisibility}
+        onToggleLayerOptions={showOpacityControl && handleToggleLayerOptions}
+        isLayerOptionsExpanded={isLayerOptionsExpanded}
       />
       {!!children && (
         <Collapse ref={wrapper} in={expanded} timeout='auto' unmountOnExit>
@@ -73,6 +93,14 @@ export default function LegendWrapper({
                 </Typography>
               )}
               {children}
+              <Collapse in={isLayerOptionsExpanded} timeout='auto' unmountOnExit>
+                <Box className={classes.layerOptionsWrapper} mt={2}>
+                  <OpacityControl
+                    opacity={opacity}
+                    onChangeOpacity={handleChangeOpacity}
+                  />
+                </Box>
+              </Collapse>
               <Note>{note}</Note>
             </Grid>
           </Box>
@@ -119,9 +147,12 @@ function Header({
   collapsible,
   expanded,
   onExpandClick,
-  onChangeVisibility
+  onChangeVisibility,
+  onToggleLayerOptions,
+  isLayerOptionsExpanded
 }) {
   const classes = useHeaderStyles({ collapsible });
+  const ExpandIcon = expanded ? ExpandLess : ExpandMore;
 
   return (
     <Grid container alignItems='center' className={classes.header}>
@@ -130,11 +161,7 @@ function Header({
         startIcon={
           collapsible && (
             <Icon>
-              {expanded ? (
-                <ExpandLess className={classes.expandIcon} />
-              ) : (
-                <ExpandMore className={classes.expandIcon} />
-              )}
+              <ExpandIcon className={classes.expandIcon} />
             </Icon>
           )
         }
@@ -142,6 +169,17 @@ function Header({
       >
         <Typography variant='subtitle1'>{title}</Typography>
       </Button>
+      {!!onToggleLayerOptions && (
+        <Tooltip title='Layer options' placement='top' arrow>
+          <ToggleButton
+            selected={isLayerOptionsExpanded}
+            onClick={onToggleLayerOptions}
+            value='check'
+          >
+            <LayerIcon />
+          </ToggleButton>
+        </Tooltip>
+      )}
       {switchable && (
         <Tooltip title={(visible ? 'Hide' : 'Show') + ' layer'} placement='top' arrow>
           <Switch checked={visible} onChange={onChangeVisibility} />